+++ apps/@examples/modal-custom/src/components/skeleton/index.tsx
@@ -1,0 +1,30 @@
+export function Skeleton() {
+  return (
+    <svg
+      viewBox="0 0 800 578"
+      fill="#343a40"
+      xmlns="http://www.w3.org/2000/svg"
+      className="skeleton"
+    >
+      <rect width="128" height="42" rx="8" />
+      <rect x="618" width="42" height="42" rx="8" />
+      <rect x="672" y="9" width="128" height="24" rx="8" />
+      <rect y="74" width="128" height="24" rx="8" />
+      <rect x="152" y="74" width="128" height="24" rx="8" />
+      <rect x="304" y="74" width="128" height="24" rx="8" />
+      <rect x="672" y="74" width="128" height="24" rx="8" />
+      <rect y="130" width="384" height="128" rx="8" />
+      <rect y="274" width="384" height="24" rx="8" />
+      <rect y="314" width="300" height="24" rx="8" />
+      <rect x="416" y="130" width="384" height="128" rx="8" />
+      <rect x="416" y="274" width="384" height="24" rx="8" />
+      <rect x="416" y="314" width="306" height="24" rx="8" />
+      <rect y="370" width="384" height="128" rx="8" />
+      <rect y="514" width="384" height="24" rx="8" />
+      <rect y="554" width="192" height="24" rx="8" />
+      <rect x="416" y="370" width="384" height="128" rx="8" />
+      <rect x="416" y="514" width="384" height="24" rx="8" />
+      <rect x="416" y="554" width="358" height="24" rx="8" />
+    </svg>
+  );
+}