+++ components/@dopt/react/modal/sandbox/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @dopt/react-modal-sandbox
 
+## 0.0.8
+
+### Patch Changes
+
+- Updated dependencies [22a71df13]
+  - @dopt/react-modal@1.0.6
+
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies [8963a5c29]
+  - @dopt/react-modal@1.0.5
+
 ## 0.0.6
 
 ### Patch Changes